added View component and hooks.tsx file
