+++ frontend/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
 import './index.css';
-import {CalendarView, HomeView, LoginView} from './views';
+import {CalendarView, HomeView, LoginPage} from './views';
 import reportWebVitals from './reportWebVitals';
 import {ProvideAuth} from './utils/auth';
 import { PrivateRoute } from './components';
@@ -13,7 +13,7 @@ ReactDOM.render(
       <Switch>
         <PrivateRoute path='/calendar' component={CalendarView} />
         <PrivateRoute path='/home' component={HomeView} />
-        <Route exact path='/' component={LoginView} />
+        <Route exact path='/' component={LoginPage} />
       </Switch>
     </Router>
   </ProvideAuth>,
