+++ frontend/src/views/index.tsx
@@ -1,3 +1,3 @@
 export {CalendarView} from './CalendarView';
 export {HomeView} from './HomeView';
-export {LoginView} from './LoginView';
+export {LoginPage} from './LoginPage';
