+++ frontend/src/views/temporaryData.tsx
@@ -1,0 +1,76 @@
+const columns = [
+    {
+      title: 'client',
+      dataIndex: 'client',
+      key: 'client',
+    },
+    {
+      title: 'project',
+      dataIndex: 'project',
+      key: 'project',
+    },
+    {
+      title: 'task',
+      dataIndex: 'task',
+      key: 'task',
+    },
+    {
+      title: 'time',
+      key: 'time',
+      dataIndex: 'time',
+    }
+  ];
+  
+  const data = [
+    {
+      key: '1',
+      client: 'KFC',
+      project: 'Kubełek',
+      task: 'kurczak',
+      time: 10,
+    },
+    {
+      key: '2',
+      client: 'KFC',
+      project: 'grander',
+      task: 'kurczak',
+      time: 60,
+    },
+    {
+      key: '3',
+      client: 'KFC',
+      project: 'bsmart',
+      task: 'wołowina',
+      time: 40,
+    },
+    {
+      key: '4',
+      client: 'KFC',
+      project: 'zinger',
+      task: 'kurczak',
+      time: 60,
+    },
+    {
+      key: '5',
+      client: 'KFC',
+      project: 'grander texas',
+      task: 'kurczak',
+      time: 50,
+    },
+    {
+      key: '6',
+      client: 'Jim Green',
+      project: 'Kubełek',
+      task: 'kurczak',
+      time: 20,
+    },
+    {
+      key: '7',
+      client: 'Joe Black',
+      project: 'Kubełek',
+      task: 'kurczak',
+      time: 10,
+    }
+  ];
+
+  export {columns, data}
